allows a strategy where each document is downloaded as soon as possible
For some backends (in particular the new approach taken by `societegenerale`) then
document urls are 'signed' and are only valid for the current 'page' of document
listing.
Meaning that when you call `iter_documents()` and it causes a fetch of a new page
of document listings, then all your document urls are invalid now.
By setting this new config option to `True`, the documents will be fetched inside
the `iter_documents()` loop, when the document url is valid.

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>