allow customization of slash character replacement in templated variables
The `filename` definition can use templated variables coming from subscription,
document and document.format (as `extension`).
A security mechanism is in place to prevent any '/' appearing in these variables
to be interpreted as a path separator ; and thus '/' are replaced by default
with `_slash_`.

This PR adds the ability to configure said replacement with any string (or char)
of your liking - with the config directive `slash_character_replacement`.

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>